--- src/constants/routes.js
@@ -1,6 +1,0 @@
-export const ADD = "/add";
-export const HOME = "/";
-export const EDIT = "/edit";
-export const EDIT_ID = "/edit/:id";
-export const DETAIL = "/p";
-export const DETAIL_ID = "/p/:id";
