--- src/views/AddForm/index.js
@@ -1,3 +1,0 @@
-import AddFrom from './AddForm';
-
-export default AddFrom;
