--- src/views/EditForm/index.js
@@ -1,3 +1,0 @@
-import EditForm from './EditForm';
-
-export default EditForm;
