+++ src/views/Layouts/index.js
@@ -1,0 +1,5 @@
+import HeaderLayout from './HeaderLayout';
+
+export {
+  HeaderLayout
+}
