--- src/views/PhoneDetail/index.js
@@ -1,3 +1,0 @@
-import PhoneDetail from './PhoneDetail';
-
-export default PhoneDetail;
