--- src/views/PhoneList/index.js
@@ -1,3 +1,0 @@
-import PhoneList from './PhoneList';
-
-export default PhoneList;
