+++ src/views/Phones/index.js
@@ -1,0 +1,11 @@
+import PhoneDetail from './PhoneDetail';
+import PhoneList from './PhoneList';
+import AddPhone from './AddPhone';
+import EditPhone from './EditPhone';
+
+export {
+  PhoneList,
+  PhoneDetail,
+  AddPhone,
+  EditPhone
+};
